Add a new "Intent" class.
This factors out a lot of "helper" glue that the `Bridge` class would otherwise
have to do. Performing by intent helps solve impedence mismatches between
networks. For example, assume a network that has no concept of joining or
inviting users to a room. Designers of the bridge for this network need not
care about invites/joins if they design by intent. They can just call
`sendMessage` and the Intent class will make sure that the user is
invited/joined correctly, as the intent was to send a message "by whatever
means necessary".