Dial a different conference ext depending on fs_user_id
Extensions loop on 35xx (00-99), and inviting the same fs_user_id gets
the same extension #. This does put a limit of 100 *concurrent* conferences
on the AS, but for now that is acceptable. Bridge: Set the matrix room/sender
from the db rather than the stub objects, as they may have important
key-value mappings attached to them which won't be obtained otherwise.