Start integrating with the target room; specify fs user ID format
An @fs_ user needs to be in the target room in order to make join/parts
in that room control who can be on the conference. Settle on the format
of @fs_$B64-ENCODED-ROOMID-MINUS-PADDING:domain for the conf user to get
around url encoding restrictions on user ID localparts. Fix a bug in the
Bridge which passed the bot Intent instead of the MatrixClient.