Change the function signature for Cli.generateRegistration
Pass an AppServiceRegistration object instead of a String. Using a
string was a bad idea because it meant that devs had to require()
matrix-appservice in order to import the AppServiceRegistration
class, violating the idea that the bridge should be encapsulating
the core appservice lib. We now pass an instantiated AppServiceRegistration
so devs don't need an extra dependency in their bridges.