Intent: Handle not in room failures due to a stale cache
It is possible that Synapse will not send join/leave `m.room.member`
events to the bridge which were originally sent **by** the bridge.
This breaks an assumption in the `Intent` class whereby we maintain
a membership cache which is kept up-to-date by calls to `onEvent`.
This cache can now get out of sync.

To fix this, we now have "join guards" around the sending of
message and state events. This guard intercepts failures and checks
if the `errcode` is `M_FORBIDDEN`. If it is, it will perform another
call to `_ensureJoined` but with the `ignoreCache` flag set, forcing
an HTTP hit to join the room.

Fixes matrix-org/matrix-appservice-irc#343